api/requirements.txt
Flask==2.0.1
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1617877307939/work
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1624781299557/work
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621523467000/work
numpy==1.21.1
pandas==1.3.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1
zipp @ file:///tmp/build/80754af9/zipp_1625570634446/work
